Basic container wireup for MVC App
Lots of MVC app cruft and only wired up for Controllers at the moment
but controllers are registered now. The API of the container is a little
weird too with the container.Register<HomeController, HomeController>().

Maybe a Register<T> method that hides that bit of awkwardness. Also look
into possibly scanning and registering all Controllers?