Port code to .NET core assemblies
Still have the old ones around for now but those will get deleted and
the new ones will get renamed soon.

Also had to make a few changes to do a GetTypeInfo on the Type class
since .NET core moved some of the APIs down to the TypeInfo. Probably
will refactor the code to pass around TypeInfo instead of Type.